signal_worker/requirements.txt
python-dotenv==1.0.1
sqlalchemy==2.0.29
asyncpg==0.30.0
pandas==2.2.2
prometheus_client==0.20.0
